+++ TvMaze/Modules/ShowDetail/View/ShowDetailView.swift
@@ -85,7 +85,7 @@ extension ShowDetailView: ShowDetailViewInterface {
         lblShowTitle.text = show.name
         lblShowLanguageValue.text = show.language
         lblShowStatusValue.text = show.status
-        setImageFrom(show.imagen.medium, showImage, UIImage(named: "show-icon")!)
+        KingFisherHelper.setImageFrom(url: show.imagen.medium, in: showImage, withPlaceholder: UIImage(named: "show-icon")!)
     }
     
     func showUIElements() {
@@ -117,8 +117,6 @@ extension ShowDetailView: UITableViewDelegate, UITableViewDataSource {
     
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(forIndexPath: indexPath) as EpisodeTableViewCell
-        let imagen = episodesViewModel[indexPath.row].imagen?.original
-        setImageFrom(imagen, cell.episodeImage, UIImage(named: "show-icon")!)
         cell.set(episode: episodesViewModel[indexPath.row])
         return cell
     }
