+++ TvMaze/Utils/Extensions/UIViewController.swift
@@ -7,28 +7,8 @@
 //
 
 import UIKit
-import Kingfisher
 
 extension UIViewController {
-    
-    func setImageFrom(_ url: String?, _ imageView: UIImageView, _ placeholder: UIImage) {
-        guard let url = URL(string: url ?? "") else {
-            imageView.image = placeholder
-            return
-        }
-        let urlResource = url
-        let processor = RoundCornerImageProcessor(cornerRadius: 20)
-        imageView.kf.indicatorType = .activity
-        imageView.kf.setImage(
-            with: urlResource,
-            placeholder: placeholder,
-            options: [
-                .processor(processor),
-                .scaleFactor(UIScreen.main.scale),
-                .transition(.fade(1)),
-                .cacheOriginalImage
-            ])
-    }
     
     func showActivityIndicator(activityIndicator: UIActivityIndicatorView) {
         activityIndicator.center = self.view.center
